[GEOT-4920] ContentDataStore.getFeatureSource(String) calls createTypeNames() every time
